% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_pvals.R
\name{approx_pvals}
\alias{approx_pvals}
\title{Approximate permutation p-values}
\usage{
approx_pvals(
  tPerm,
  tObs,
  tol = 1e-08,
  eps = NULL,
  includeObs = TRUE,
  fitMethod = "MLE",
  finalOnly = FALSE,
  constraint = "tObs",
  fitThresh = 0.1,
  nonZero = TRUE,
  appMethod = "gpd",
  threshVec = NULL,
  threshMethod = "PRbelowAlpha",
  thresh0 = NULL,
  nExceed0 = NULL,
  nExceedMin = 1,
  stepSize = 1,
  gofTest = "ad",
  gofAlpha = 0.05,
  gofTailRMMeth = "allrej",
  gofTailRMPar = NULL,
  seed = NULL,
  cores = 1L,
  verbose = FALSE,
  gpd_estimate = NULL,
  plotHist = F,
  histBreaks = 100,
  histXlim = NULL,
  histYlim = NULL,
  plotPvals = TRUE,
  plotPvalsxvar = "thresh",
  plotTitle = "",
  maintext = ""
)
}
\arguments{
\item{tPerm}{numeric vector with permutation test statistics}

\item{tObs}{numeric value giving the observed test statistic}

\item{includeObs}{logical. Indicates whether the observed test statistic
should be included in the permutation distribution.}

\item{fitMethod}{...}

\item{fitThresh}{threshold for initial p-value, above which a GPD or Gamma
distribution is fitted for p-value approximation (e.g. 2 times alpha).
Defaults is 0.1.}

\item{nonZero}{logical. If \code{TRUE} (default), a p-value of zero is
strictly avoided.}

\item{appMethod}{character indicating the method used for p-value estimation.
Possible values are: "gpd" (default), "gamma", "empirical".}

\item{threshMethod}{method for threshold detection.
Possible values are:
"fix", "ftr", "minPR", "PRbelowAlpha", "fwdStop"}

\item{nExceedMin}{minimum number of exceedances}

\item{stepSize}{numeric giving the value by which either the threshold or the
number of exceedances is increased in each step, depending on threshMethod}

\item{gofTest}{goodness-of-fit test method}

\item{gofAlpha}{significance level of the GOF test}

\item{seed}{...}

\item{gpd_estimate}{...}

\item{threshMethodPar}{fix threshold/number of exceedances or start value for
threshold / number of exceedances, depending on threshMethod}

\item{ffwd}{if TRUE, large steps are made at the beginning until H0 is accepted
for the first time}

\item{fitInformation}{...}

\item{optimMethod}{...}

\item{shape0, scale0}{start value of the shape and scale parameter}

\item{shapeMin, scaleMin}{minimum value of the shape and scale parameter}

\item{shapeMax, scaleMax}{maximum value of the shape and scale parameter}
}
\description{
Approximate the p-value for a one-sided permutation hypothesis test,
  where the p-value is defined as the probability of observing a test
  statistic (in absolute terms) greater than or equal to the observed value
  under the null. Approximation methods include the empirical p-value,
  Gamma approximation, and GPD tail approximation.
}
