% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_utilities.R
\name{.choose_workers}
\alias{.choose_workers}
\title{Define number of workers for parallel processing}
\usage{
.choose_workers(cores, n, parallel_min)
}
\description{
Define number of workers for parallel processing
}
\keyword{internal}
