% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_compute_pvals_emp.R
\name{.compute_pvals_emp}
\alias{.compute_pvals_emp}
\title{Compute empirical permutation p-values}
\usage{
.compute_pvals_emp(obs_stats, perm_stats)
}
\description{
Compute empirical permutation p-values
}
\keyword{internal}
