% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_compute_pvals_gpd.R
\name{.compute_pvals_gpd}
\alias{.compute_pvals_gpd}
\title{Compute p-values via GPD tail approximation}
\usage{
.compute_pvals_gpd(
  p_empirical,
  obs_stats,
  perm_stats,
  n_test,
  fit_thresh,
  alternative,
  control,
  ...
)
}
\description{
Compute p-values via GPD tail approximation
}
\keyword{internal}
