% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps_functions.R
\name{.define_eps}
\alias{.define_eps}
\title{Define epsilon values based on the user-defined rule}
\usage{
.define_eps(
  perm_stats,
  obs_stats,
  sample_size,
  constraint,
  eps_fun,
  eps_tune,
  eps_args
)
}
\description{
Define epsilon values based on the user-defined rule
}
\keyword{internal}
