% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_find_gpd_thresh.R
\name{.find_gpd_thresh}
\alias{.find_gpd_thresh}
\title{Adapted version of gpdAd from eva package}
\usage{
.find_gpd_thresh(
  perm_stats,
  obs_stats,
  tol,
  thresh_method = "PRbelowAlpha",
  thresh0 = NULL,
  exceed0 = NULL,
  exceed_min = 1,
  thresh_step = 1,
  gof_test = "ad",
  gof_alpha = 0.05,
  seed = NULL,
  doPlot = FALSE,
  ...
)
}
\description{
Adapted version of gpdAd from eva package
}
\keyword{internal}
