% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_gpd_fit_methods.R
\name{.fit_gpd_lme}
\alias{.fit_gpd_lme}
\title{Likelihood Moment Estimation for the GPD}
\usage{
.fit_gpd_lme(x, boundary = NULL, eval_point = NULL, r = -1/2, tol = 1e-08)
}
\arguments{
\item{x}{data vector}

\item{boundary}{numeric. Maximum value at which the GPD probability density
function must be positive. If \code{NULL}, the maximum of the data vector
\code{x} is used.}

\item{r}{numeric defining the r constant as part of the algorithm that must
be smaller than 1/2 and different from zero. Default is -1/2.
See details.}

\item{tol}{tolerance used as convergence criterion. The iteration stops if
|f(b)/f'(b)/b| < tol. Defaults to 1e-8.}
}
\description{
Likelihood Moment Estimation for the two-parameter
  Generalized Pareto Distribution (GPD) proposed by \cite{Zhang (2007)}.
}
\details{
According to \cite{Zhang (2007)}, the constant r must be smaller than 1/2
  and different from zero (r < 1/2 and r != 0).
  r close to -shape leads to estimates close to the MLEs.
  Without knowledge about the true shape, r = -1/2 (default)
  "produces LMEs with high asymptotic efficiency" \cite{(Zhang 2007)}.
  \cr\cr
}
\references{
\insertRef{Zhang2007lme}{permAprox}
}
\keyword{internal}
