% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_gpd_fit_methods.R
\name{.fit_gpd_mle1d}
\alias{.fit_gpd_mle1d}
\title{One-dimensional MLE for the GPD}
\usage{
.fit_gpd_mle1d(
  x,
  boundary = NULL,
  eval_point = NULL,
  shapePos = FALSE,
  tol = 1e-08
)
}
\arguments{
\item{x}{data vector}

\item{boundary}{numeric. Maximum value at which the GPD probability density
function must be positive. If \code{NULL}, the maximum of the data vector
\code{x} is used.}

\item{shapePos}{logical. If \code{TRUE}, the optimization is done under the
constraint of a positive shape parameter.}

\item{tol}{numeric giving the desired accuracy of the optimization
process.}
}
\description{
One-dimensional Maximum Likelihood Estimation for the two-parameter
  Generalized Pareto Distribution (GPD) proposed by
  \cite{Castillo & Serra (2015)}.
}
\references{
\insertRef{Castillo2015likelihood}{permAprox}
}
\keyword{internal}
