% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_gpd_fit_methods.R
\name{.fit_gpd_nls2}
\alias{.fit_gpd_nls2}
\title{Parameter estimation via two-step NLS for the GPD}
\usage{
.fit_gpd_nls2(
  x,
  q = 0,
  boundary = NULL,
  eval_point = NULL,
  twosteps = TRUE,
  optimMethod = "Nelder-Mead",
  tol = 1e-09,
  shapeIni = NULL,
  scaleIni = NULL,
  shapeMin = -Inf,
  scaleMin = -Inf,
  shapeMax = Inf,
  scaleMax = Inf,
  ...
)
}
\arguments{
\item{x}{data vector}

\item{q}{threshold value. E.g., set to q=0.9 to use the 90% percentile.
Default is 0, i.e., all data are used.}

\item{boundary}{numeric. Maximum value at which the GPD probability density
function must be positive. If \code{NULL}, the maximum of the data vector
\code{x} is used.}

\item{tol}{numeric giving the desired accuracy of the optimization
process.}

\item{shapeIni, scaleIni}{initial values for the shape and scale parameters,
i.e., values where the optimization starts.}

\item{shapeMin, scaleMin}{lower bound on the shape and scale parameters.
If != -Inf, \code{optimMethod} is set to "L-BFGS-B".}

\item{shapeMax, scaleMax}{upper bound on the shape and scale parameters.
If != Inf, \code{optimMethod} is set to "L-BFGS-B".}
}
\description{
Parameter estimation for the two-parameter Generalized Pareto Distribution
  (GPD) using a two-step Nonlinear Least Squares (NLS) approach proposed by
  \cite{Song & Song (2012)}.
}
\references{
\insertRef{Song2012quantile}{permApprox}
}
\keyword{internal}
