% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_gpd_fit_methods.R
\name{.fit_gpd_zse}
\alias{.fit_gpd_zse}
\title{Parameter estimation for the GPD proposed by Zhang & Stephens}
\usage{
.fit_gpd_zse(x, boundary = NULL, eval_point = NULL, m = NULL)
}
\arguments{
\item{x}{data vector}

\item{boundary}{numeric. Maximum value at which the GPD probability density
function must be positive. If \code{NULL}, the maximum of the data vector
\code{x} is used.}
}
\description{
Parameter estimation for the two-parameter Generalized Pareto Distribution
  (GPD) proposed by \cite{Zhang & Stephens (2009)}.
}
\references{
\insertRef{Zhang2009new}{permAprox}
}
\keyword{internal}
