% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_gof_tests.R
\name{.gof_gpd_cvm}
\alias{.gof_gpd_cvm}
\title{Adapted version of gpdCvm from eva package}
\usage{
.gof_gpd_cvm(
  data,
  scale,
  shape,
  bootstrap = FALSE,
  bootnum = NULL,
  allowParallel = FALSE,
  numCores = 1
)
}
\description{
Adapted version of gpdCvm from eva package
}
\keyword{internal}
