% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_utilities.R
\name{.transform_stats}
\alias{.transform_stats}
\title{Transform test statistics for tail modeling}
\usage{
.transform_stats(perm_stats, obs_stat, alternative)
}
\description{
Transforms observed and permutation test statistics according to the specified
alternative hypothesis. For one-sided tests, this involves sign-flipping to
ensure that values to be fitted (e.g., to a Gamma or GPD) are positive and lie
in the appropriate tail of the null distribution.
}
\keyword{internal}
