% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_gof_tests.R
\name{findthresh}
\alias{findthresh}
\title{Copied helper function from the eva package}
\usage{
findthresh(data, ne)
}
\description{
Copied from the eva package (GPL-2 | GPL-3), originally
  authored by [Brian Bader [aut, cre], Jun Yan [ctb]].
See: https://cran.r-project.org/web/packages/eva/index.html
}
\keyword{internal}
