% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvals_gpd.R
\name{get_pvals_gpd}
\alias{get_pvals_gpd}
\title{Compute p-values via GPD approximation}
\usage{
get_pvals_gpd(
  tObs,
  tPerm,
  nTest,
  nPerm,
  ntPerm,
  useAllPerm,
  constraint,
  fitThresh,
  includeObs,
  fitMethod,
  tol,
  eps,
  epsVal,
  threshMethod,
  thresh0,
  nExceed0,
  nExceedMin,
  stepSize,
  gofTest,
  gofAlpha,
  gofTailRMMeth,
  gofTailRMPar,
  cores,
  verbose,
  ...
)
}
\description{
Compute p-values via GPD approximation
}
\keyword{internal}
