% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_MLE2D.R
\name{gpd_MLE2D}
\alias{gpd_MLE2D}
\title{Two-dimensional MLE for the GPD}
\usage{
gpd_MLE2D(
  x,
  maxX = NULL,
  eps = 0,
  optimMethod = "Nelder-Mead",
  tol = 1e-08,
  shapeIni = NULL,
  scaleIni = NULL,
  shapeMin = -Inf,
  scaleMin = -Inf,
  shapeMax = Inf,
  scaleMax = Inf,
  ...
)
}
\arguments{
\item{x}{data vector}

\item{maxX}{numeric. Maximum value at which the GPD probability density
function must be positive. If \code{NULL}, the maximum of the data vector
\code{x} is used.}

\item{eps}{numeric. Optional small epsilon, which is added to \code{maxX}
so that the GPD density is non-zero at maxX + eps.}

\item{optimMethod}{character defining the optimization method.}

\item{tol}{numeric giving the desired accuracy of the optimization
process.}

\item{shapeIni, scaleIni}{initial values for the shape and scale parameters,
i.e., values where the optimization starts.}

\item{shapeMin, scaleMin}{lower bound on the shape and scale parameters.
If != -Inf, \code{optimMethod} is set to "L-BFGS-B".}

\item{shapeMax, scaleMax}{upper bound on the shape and scale parameters.
If != Inf, \code{optimMethod} is set to "L-BFGS-B".}

\item{...}{Further arguments passed to \code{optim()}.}
}
\description{
Two-dimensional Maximum Likelihood Estimation for the two-parameter
  Generalized Pareto Distribution (GPD).
}
