% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_MOM.R
\name{gpd_MOM}
\alias{gpd_MOM}
\title{Method of Moments estimation for the GPD}
\usage{
gpd_MOM(x)
}
\arguments{
\item{x}{data vector}
}
\description{
Method of Moments estimation for the two-parameter Generalized Pareto
  Distribution (GPD) proposed by \cite{Hosking & Wallis (1987)}.
}
\references{
\insertRef{Hosking1987parameter}{permpap}
}
