% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_WNLLSM.R
\name{gpd_WNLLSM}
\alias{gpd_WNLLSM}
\title{WNLSM and WNLLSM parameter estimation for the GPD}
\usage{
gpd_WNLLSM(
  x,
  maxX = NULL,
  eps = 0,
  method = "WNLLSM",
  tol = 1e-08,
  shapeIni = 0.01,
  scaleIni = 0.01,
  F0 = 0,
  plotopt = FALSE,
  ...
)
}
\arguments{
\item{x}{data vector}

\item{maxX}{numeric. Maximum value at which the GPD probability density
function must be positive. If \code{NULL}, the maximum of the data vector
\code{x} is used.}

\item{eps}{numeric. Optional small epsilon, which is added to \code{maxX}
so that the GPD density is non-zero at maxX + eps.
\code{eps} must be larger than zero.}

\item{tol}{numeric giving the desired accuracy of the optimization
process.}

\item{shapeIni, scaleIni}{initial values for the shape and scale parameters,
i.e., values where the optimization starts.}

\item{q}{threshold value. E.g., set to q=0.9 to use the 90% percentile.
Default is 0, i.e., all data are used.}

\item{shapeMin, scaleMin}{lower bound on the shape and scale parameters.
If != -Inf, \code{optimMethod} is set to "L-BFGS-B".}

\item{shapeMax, scaleMax}{upper bound on the shape and scale parameters.
If != Inf, \code{optimMethod} is set to "L-BFGS-B".}
}
\description{
Weighted nonlinear least squares moments (WNLSM) estimation and
  likelihood WNLSM (WNLLSM) estimation for the two-parameter Generalized
  Pareto Distribution (GPD) proposed by \cite{Zhao et al. (2019)}.
}
\details{
A simplified version of the R code has been provided by
  \cite{Zhao et al. (2019)}. The function has been extended by the arguments:
  \code{maxX}, \code{eps}, and \code{tol} to ensure that the GPD
  density is non-zero at maxX + eps or max(x) + eps.
}
\references{
\insertRef{Zhao2019new}{permpap}
}
