% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_ZSE.R
\name{gpd_ZSE}
\alias{gpd_ZSE}
\title{Parameter estimation for the GPD proposed by Zhang & Stephens}
\usage{
gpd_ZSE(x, maxX = NULL, maxXOrig = NULL, m = NULL)
}
\arguments{
\item{x}{data vector}

\item{maxX}{numeric. Maximum value at which the GPD probability density
function must be positive. If \code{NULL}, the maximum of the data vector
\code{x} is used.}

\item{q}{threshold value. E.g., set to q=0.9 to use the 90% percentile.
Default is 0, i.e., all data are used.}

\item{eps}{numeric. Optional small epsilon, which is added to \code{maxX}
so that the GPD density is non-zero at maxX + eps.}

\item{optimTol}{numeric giving the desired accuracy of the optimization
process.}

\item{shapeIni, scaleIni}{initial values for the shape and scale parameters,
i.e., values where the optimization starts.}

\item{shapeMin, scaleMin}{lower bound on the shape and scale parameters.
If != -Inf, \code{optimMethod} is set to "L-BFGS-B".}

\item{shapeMax, scaleMax}{upper bound on the shape and scale parameters.
If != Inf, \code{optimMethod} is set to "L-BFGS-B".}
}
\description{
Parameter estimation for the two-parameter Generalized Pareto Distribution
  (GPD) proposed by \cite{Zhang & Stephens (2009)}.
}
\references{
\insertRef{Zhang2009new}{permAprox}
}
