% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_adjust.R
\name{mult_adjust}
\alias{mult_adjust}
\title{Multiple testing adjustment}
\usage{
mult_adjust(
  p_values,
  method = "BH",
  true_null_method = "convest",
  p_true_null = NULL,
  seq_length = 100,
  perm_stats = NULL,
  p_perm = NULL,
  cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{p_values}{Numeric. Vector with p-values.}

\item{method}{Character. Method for p-value adjustment. Options include:
\itemize{
  \item \code{"lfdr"}: local false discovery rates via \code{fdrtool}.
  \item \code{"adaptBH"}: adaptive Benjamini-Hochberg (requires estimation
  of the proportion of true nulls).
  \item Any method supported by \code{stats::p.adjust} (e.g., "holm",
  "BH", "BY").
}}

\item{true_null_method}{Character. Method to estimate the proportion of true
null hypotheses when \code{method = "adaptBH"}. Options (for
\code{limma::propTrueNull}):
\itemize{
  \item \code{"convest"} (default)
  \item \code{"lfdr"}
  \item \code{"mean"}
  \item \code{"hist"}
  \item \code{"farco"}: Farcomeni (2007) iterative plug-in method.
}}

\item{p_true_null}{Numeric or NULL. Pre-specified proportion of true nulls
for \code{adaptBH}. If \code{NULL}, it is estimated using
\code{true_null_method}. Default: \code{NULL}.}

\item{seq_length}{Integer. Number of grid points when computing FDR curves in
\code{"rbFDR"} method. Default: \code{100}.}

\item{perm_stats}{Permutation test statistics.}

\item{p_perm}{Permutation p-values.}

\item{cores}{Integer. Number of CPU cores to use for parallel computations
(e.g., in \code{"rbFDR"}). Must be >=1. Default: \code{1}.}

\item{verbose}{Logical. If \code{TRUE}, progress messages are printed.
Default: \code{FALSE}.}
}
\description{
The functions adjusts a vector of p-values for multiple testing.
}
