% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permAprox-package.R
\docType{package}
\name{permAprox-package}
\alias{permAprox-package}
\title{permAprox: Permutation p-value approximation}
\description{
Approximate p-values in permutation tests using the Generalized Pareto
Distribution (GPD) while strictly avoiding p-values of zero.\cr\cr
The main function for approximating permutation p-values is
\code{\link{permaprox}}.
}
\author{
\strong{Maintainer:} Stefanie Peschel \email{stefanie.peschel@mail.de}\cr\cr
  \strong{Acknowledgments:}\cr
  \itemize{
   \item Christian L. Müller, Anne-Laure Boulesteix
  }
}
