% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permaprox.R
\name{permaprox}
\alias{permaprox}
\title{Approximate permutation p-values}
\usage{
permaprox(
  obs_stats,
  perm_stats,
  alternative = "twoSided",
  nullValue = 0,
  method = "gpd",
  fit_thresh = 0.2,
  gammaOnFail = TRUE,
  includeObs = TRUE,
  fit_method = "MLE1D",
  constraint = "obs_statsMax",
  tol = 1e-08,
  eps = 0.8,
  eps_type = "quantile",
  thresh_method = "PRbelowAlpha",
  thresh0 = NULL,
  threshPoss = NULL,
  exceed0 = NULL,
  exceed_min = 0.1,
  thresh_step = 1,
  gof_test = "ad",
  gof_alpha = 0.05,
  gof_testGamma = TRUE,
  multAdj = "adaptBH",
  alpha = 0.05,
  true_null_method = "convest",
  p_true_null = NULL,
  seed = NULL,
  cores = 1L,
  verbose = FALSE,
  gpdEstimate = NULL,
  seq_length = 100,
  p_perm = NULL,
  ...
)
}
\arguments{
\item{obs_stats}{numeric vector giving the observed test statistic(s).}

\item{perm_stats}{numeric vector or matrix with permutation test statistics. In
the multiple testing setting, \code{perm_stats} must be a matrix where each row
contains the permutation test statistics of a test and thus belongs to an
entry in \code{obs_stats}.}

\item{method}{character indicating the method used for p-value approximation.
Possible values are: "gpd" (default; fitting the GPD distribution),
"gamma" (fitting a Gamma distribution), "empirical" (use the empirical
p-value).}

\item{fit_thresh}{p-value threshold above which a GPD or gamma distribution is
fitted to approximate the p-value (e.g., 2 times alpha). Defaults to 0.1.}

\item{includeObs}{logical. Indicates whether the observed test statistic
**obs_stats** should be included in the fitting process. Note that **obs_stats** is
always included when computing the empirical p-value.}

\item{fit_method}{character giving the method used for fitting the GPD
parameters. Possible methods are: "LME", "MLE1D", "MLE2D", "MOM", "NLS2",
"WNLLSM", and "ZSE".}

\item{constraint}{character defining the constraint, under which the GPD is
fitted. Possible values are:
\describe{
\item{\code{"none"}}{No constraint.}
\item{\code{"shapePos"}}{GPD shape parameter must be positive.}
\item{\code{"obs_stats"}}{The upper bound of f(x) (=-scale/shape) must be
greater than **obs_stats**.}
\item{\code{"obs_statsMax"}}{The upper bound of f(x) (=-scale/shape) must be
greater than the maximum **obs_stats** of all tests (in the multiple testing
setting).}
}}

\item{tol}{convergence tolerance (used for fitting GPD parameters)}

\item{eps}{small numeric value or proportion (used for constraint)}

\item{eps_type}{character defining the type of epsilon. Default is "quantile".}

\item{thresh_method}{method for threshold detection.
Possible values are: "fix", "ftr", "minPR", "PRbelowAlpha", "fwdStop"}

\item{thresh0}{numeric value giving the initial threshold.}

\item{exceed0}{numeric value giving the initial number of exceedances. Either
\code{thresh0} or \code{exceed0} must be given.}

\item{exceed_min}{numeric giving the minimum number of exceedances.}

\item{thresh_step}{numeric giving the value by which either the threshold or the
number of exceedances is increased in each step, depending on thresh_method.}

\item{gof_test}{goodness-of-fit test method. Possible values are "ad"
(Anderson-Darling-Test; default) or "cvm" (Cramér–von Mises criterion).}

\item{gof_alpha}{significance level of the GOF test.}

\item{multAdj}{character specifying the method used to adjust for multiple
testing. Possible methods are "lfdr" (local FDR), "adaptBH" (default;
adaptive Benjamini Hochberg), "rbFDR" (resampling-based FDR), as well as
the methods contained in \code{p.adjust.methods}.}

\item{alpha}{significance level.}

\item{true_null_method}{method used for identifying the proportion of
true null hypotheses. Only used if \code{multAdj} is set to "adaptBH".}

\item{p_true_null}{proportion of true null hypotheses.}

\item{seed}{random seed.}

\item{cores}{number of cores used for parallelization.}

\item{verbose}{logical. If TRUE, messages returned by internal functions are
printed.}

\item{gpdEstimate}{not used.}

\item{seq_length}{length of the sequence of possible cut points. Used for
resampling-based multiple testing adjustment.}

\item{p_perm}{description}
}
\description{
Approximate p-values of one-tailed permutation hypothesis tests, where the
 p-value is defined as the probability that a test statistic (in absolute
 terms) is greater than or equal to the observed value under the null.
 Approximation methods include the empirical p-value, the gamma
 approximation, and the GPD tail approximation.
}
