% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/.pgpd_upper_tail.R
\name{.pgpd_upper_tail}
\alias{.pgpd_upper_tail}
\title{Quantile function for the upper tail of the GPD}
\usage{
.pgpd_upper_tail(q, location = 0, shape, scale)
}
\arguments{
\item{q}{quantile}

\item{location}{location parameter of the GPD}

\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\description{
Enables the computation of very small p-values.
}
