% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_perm_approx.R
\name{print.perm_approx}
\alias{print.perm_approx}
\title{Print method for permApprox results}
\usage{
\method{print}{perm_approx}(x, digits = 3L, ...)
}
\arguments{
\item{x}{An object of class \code{"perm_approx"}.}

\item{digits}{Number of digits for numeric summaries.}

\item{...}{Ignored.}
}
\description{
Print method for permApprox results
}
