% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_compute_pvals_gamma.R
\name{.compute_pvals_gamma}
\alias{.compute_pvals_gamma}
\title{Compute p-values via Gamma approximation}
\usage{
.compute_pvals_gamma(
  p_empirical,
  perm_stats,
  obs_stats,
  n_test,
  fit_thresh,
  alternative,
  control
)
}
\description{
Compute p-values via Gamma approximation
}
\keyword{internal}
