% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gpd.R
\name{fit_gpd}
\alias{fit_gpd}
\title{Fit and test the GPD}
\usage{
fit_gpd(
  data,
  thresh = NULL,
  fitMethod = "MLE1D",
  tol = 1e-08,
  eps = NULL,
  epsVal = 0.9,
  constraint = "none",
  maxVal = NULL,
  gofTest = "ad",
  ...
)
}
\arguments{
\item{data}{data vector}

\item{thresh}{numeric defining the GPD threshold. Data above this threshold
are the exceedances.}

\item{fitMethod}{!!! character giving the method used for fitting the GPD
distribution.}

\item{maxVal}{maximum value for which the probability density function must
be positive. If \code{NULL}, the maximum of the data vector is used.}

\item{...}{arguments passed to the function for GPD parameter estimation.}

\item{nExceed}{integer giving the number of exceedances (data above the
threshold).}
}
\description{
Fit and test the GPD
}
\details{
The maximum value (\code{maxVal}) is only relevant for light-tailed
distributions, i.e. if the shape parameter is negative.
}
