% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvals_emp.R
\name{get_pvals_emp}
\alias{get_pvals_emp}
\title{Compute empirical permutation p-values}
\usage{
get_pvals_emp(tObs, tPerm, nTest, nPerm, ntPerm, useAllPerm)
}
\description{
Compute empirical permutation p-values
}
\keyword{internal}
