% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvals_gamma.R
\name{get_pvals_gamma}
\alias{get_pvals_gamma}
\title{Compute p-values via Gamma approximation}
\usage{
get_pvals_gamma(
  p_empirical,
  perm_stats,
  obs_stats,
  n_test,
  fit_thresh,
  alternative,
  gof_testGamma,
  gof_alpha
)
}
\description{
Compute p-values via Gamma approximation
}
\keyword{internal}
