% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvals_gamma.R
\name{get_pvals_gamma}
\alias{get_pvals_gamma}
\title{Compute p-values via Gamma approximation}
\usage{
get_pvals_gamma(pvals, tPerm, tObs, nTest, fitThresh)
}
\description{
Compute p-values via Gamma approximation
}
\keyword{internal}
