% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvals_gpd.R
\name{get_pvals_gpd}
\alias{get_pvals_gpd}
\title{Compute p-values via GPD approximation}
\usage{
get_pvals_gpd(
  tObs,
  tPerm,
  alternative,
  pEmp,
  nExtreme,
  nTest,
  nPerm,
  constraint,
  fitThresh,
  gammaOnFail,
  includeObs,
  fitMethod,
  tol,
  eps,
  epsType,
  threshMethod,
  thresh0,
  threshPoss,
  exceed0,
  exceedMin,
  stepSize,
  gofTest,
  gofAlpha,
  cores,
  verbose,
  ...
)
}
\description{
Compute p-values via GPD approximation
}
\keyword{internal}
