% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvals_gpd.R
\name{get_pvals_gpd}
\alias{get_pvals_gpd}
\title{Compute p-values via GPD approximation}
\usage{
get_pvals_gpd(
  obs_stats,
  perm_stats,
  alternative,
  pEmp,
  nExtreme,
  nTest,
  nPerm,
  constraint,
  fit_thresh,
  gammaOnFail,
  gof_testGamma,
  includeObs,
  fit_method,
  tol,
  eps,
  eps_type,
  thresh_method,
  thresh0,
  threshPoss,
  exceed0,
  exceed_min,
  thresh_step,
  gof_test,
  gof_alpha,
  cores,
  verbose,
  ...
)
}
\description{
Compute p-values via GPD approximation
}
\keyword{internal}
