% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_MLE1D.R
\name{gpd_MLE1D}
\alias{gpd_MLE1D}
\title{One-dimensional MLE for the GPD}
\usage{
gpd_MLE1D(x, maxX = NULL, shapePos = FALSE, optimTol = 1e-08, eps = 0)
}
\arguments{
\item{x}{data vector}

\item{maxX}{numeric. Maximum value at which the GPD probability density
function must be positive. If \code{NULL}, the maximum of the data vector
\code{x} is used.}

\item{shapePos}{logical. If \code{TRUE}, the optimization is done under the
constraint of a positive shape parameter.}

\item{optimTol}{numeric giving the desired accuracy of the optimization
process.}

\item{eps}{numeric. Optional small epsilon, which is added to \code{maxX}
so that the GPD density is non-zero at maxX + eps.}
}
\description{
One-dimensional Maximum Likelihood Estimation for the two-parameter
  Generalized Pareto Distribution (GPD) proposed by
  \cite{Castillo & Serra (2015)}.
}
\references{
\insertRef{Castillo2015likelihood}{permpap}
}
