% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/.fit_gpd_mom.R
\name{.fit_gpd_mom}
\alias{.fit_gpd_mom}
\title{Method of Moments estimation for the GPD}
\usage{
.fit_gpd_mom(x)
}
\arguments{
\item{x}{data vector}
}
\description{
Method of Moments estimation for the two-parameter Generalized Pareto
  Distribution (GPD) proposed by \cite{Hosking & Wallis (1987)}.
}
\references{
\insertRef{Hosking1987parameter}{permAprox}
}
