% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adjust_ctrl.R
\name{make_adjust_ctrl}
\alias{make_adjust_ctrl}
\title{Create a control object for multiple testing adjustment in permAprox}
\usage{
make_adjust_ctrl(
  method = "adaptBH",
  true_null_method = "convest",
  p_true_null = NULL,
  seq_length = 100,
  cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{method}{Character. Method for p-value adjustment. Options include:
\itemize{
  \item \code{"lfdr"}: local false discovery rates via \code{fdrtool}.
  \item \code{"adaptBH"}: adaptive Benjamini-Hochberg (requires estimation of the proportion of true nulls).
  \item Any method supported by \code{stats::p.adjust} (e.g., "holm", "BH", "BY").
}}

\item{true_null_method}{Character. Method to estimate the proportion of true null hypotheses
when \code{method = "adaptBH"}. Options (for \code{limma::propTrueNull}):
\itemize{
  \item \code{"convest"} (default)
  \item \code{"lfdr"}
  \item \code{"mean"}
  \item \code{"hist"}
  \item \code{"farco"}: Farcomeni (2007) iterative plug-in method.
}}

\item{p_true_null}{Numeric or NULL. Pre-specified proportion of true nulls for \code{adaptBH}.
If \code{NULL}, it is estimated using \code{true_null_method}. Default: \code{NULL}.}

\item{seq_length}{Integer. Number of grid points when computing FDR curves in
\code{"rbFDR"} method. Default: \code{100}.}

\item{cores}{Integer. Number of CPU cores to use for parallel computations
(e.g., in \code{"rbFDR"}). Must be >=1. Default: \code{1}.}

\item{verbose}{Logical. If \code{TRUE}, progress messages are printed.
Default: \code{FALSE}.}
}
\value{
A named list (class \code{"adjust_ctrl"}) with the specified settings.
}
\description{
Constructs a list of parameters controlling multiple testing correction
methods, so that these options are decoupled from the main
\code{compute_p_values()} function.
}
