% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adjust_ctrl.R
\name{make_adjust_ctrl}
\alias{make_adjust_ctrl}
\title{Create a control object for multiple testing adjustment in permApprox}
\usage{
make_adjust_ctrl(true_null_method = "convest", p_true_null = NULL)
}
\arguments{
\item{true_null_method}{Character. Method to estimate the proportion of true
null hypotheses when \code{method = "adapt_BH"}. Options (for
\code{limma::propTrueNull}):
\itemize{
  \item \code{"convest"} (default)
  \item \code{"lfdr"}
  \item \code{"mean"}
  \item \code{"hist"}
  \item \code{"farco"}: Farcomeni (2007) iterative plug-in method.
}}

\item{p_true_null}{Numeric or NULL. Pre-specified proportion of true nulls
for \code{adapt_BH}. If \code{NULL}, it is estimated using
\code{true_null_method}. Default: \code{NULL}.}
}
\value{
A named list (class \code{"adjust_ctrl"}) with the specified settings.
}
\description{
Constructs a list of parameters controlling multiple testing correction
methods, so that these options are decoupled from the main
\code{perm_approx()} function.
}
