% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gpd_ctrl.R
\name{make_gpd_ctrl}
\alias{make_gpd_ctrl}
\title{Create a control list for GPD approximation}
\usage{
make_gpd_ctrl(
  fit_method = "MLE1D",
  include_obs = FALSE,
  constraint = "unconstrained",
  eps = 0.8,
  eps_type = "quantile",
  tol = 1e-08,
  thresh_method = "fix",
  thresh0 = NULL,
  thresh_step = 1,
  exceed0 = 0.25,
  exceed_min = 10,
  gof_test = "ad",
  gof_alpha = 0.05,
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{fit_method}{Character. Method for GPD fitting. Options: "LME", "MLE1D",
"MLE2D", "MOM", "NLS2", "WNLLSM", "ZSE". Default: "MLE1D".}

\item{include_obs}{Logical. If \code{TRUE}, the observed test statistic is
included into the tail approximation (considered as permutation test
statistic). Default: \code{FALSE}.}

\item{constraint}{Character. Constraint for the fitting process.
Default: "unconstrained". Options:
\describe{
  \item{\code{"unconstrained"}}{No constraint.}
  \item{\code{"shape_nonneg"}}{Shape parameter must be non-negative.}
  \item{\code{"support_at_obs"}}{Positive density for observed test statistic.}
  \item{\code{"support_at_max"}}{Positive density for maximum of all
  observed test statistics (in the multiple testing case).}
}}

\item{eps}{Numeric. Small value or proportion (used for constraint)}

\item{eps_type}{Character. Defines the type of epsilon. Options: "quantile",
"fix". Default: "quantile".}

\item{tol}{Numeric. Convergence tolerance for fitting GPD parameters.
Default: 1e-8.}

\item{thresh_method}{Character. Method for threshold detection.
Default: "fix". Options:
\describe{
  \item{\code{"fix"}}{Fix threshold (defined via \code{thresh0}).}
  \item{\code{"ftr"}}{Failure to reject.}
  \item{\code{"ftr_min5"}}{Failure to reject (min 5 subsequent acceptances).}
  \item{\code{"pr_below_alpha"}}{Proportion of rejections below alpha.}
  \item{\code{"fwd_stop"}}{Forward Stop}
  \item{\code{"gof_cp"}}{GOF change point}
}}

\item{thresh0}{Numeric or NULL. Initial threshold for defining exceedances
in the GPD fitting. Is used as fixed threshold when
\code{thresh_method = "fix"}. Set to \code{NULL} if \code{exceed0} is
provided instead. Default is \code{NULL}.
Either \code{thresh0} or \code{exceed0} must be specified (non-NULL).}

\item{thresh_step}{Integer. Step size for adaptive threshold search.
If \code{thresh_step} is 2, for instance, every second possible threshold
is considered to save runtime.}

\item{exceed0}{Numeric or NULL. Initial number of exceedances.
If less than 1, it is interpreted as a proportion of the number of
permutations. Default is \code{0.25}. Either \code{exceed0} or
\code{thresh0} must be specified (non-NULL).}

\item{exceed_min}{Integer. Minimum exceedances needed for fitting. Default: 10.}

\item{gof_test}{Character. Goodness-of-fit test for GPD. Options:
"ad" (Anderson-Darling), "cvm" (Cramer-von-Mises), "none". Default: "ad".}

\item{gof_alpha}{Numeric. Significance level for GOF test (0 < gof_alpha < 1).
Default: 0.05.}

\item{cores}{Integer. Number of CPU cores to use for parallel computations.
Must be >=1. Default: \code{1}.}

\item{verbose}{Logical. If \code{TRUE}, progress messages are printed.
Default: \code{TRUE}.}
}
\value{
A named list of class "gpd_ctrl" containing GPD settings.
}
\description{
Constructs a list of control parameters for the Generalized Pareto
Distribution (GPD) tail approximation method used in permApprox.
}
