% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_adjust.R
\name{mult_adjust}
\alias{mult_adjust}
\title{Multiple testing adjustment}
\usage{
mult_adjust(
  pvals,
  perm_stats = NULL,
  p_perm = NULL,
  adjust = "adaptBH",
  true_null_method = "convest",
  p_true_null = NULL,
  seq_length = 100,
  cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{pvals}{numeric vector with p-values}

\item{adjust}{character specifying the method used for adjustment.
Can be \code{"lfdr"}, \code{"adaptBH"}, or one of the methods provided by
\code{\link[stats]{p.adjust}}.}

\item{true_null_method}{character indicating the method used for estimating the
proportion of true null hypotheses from a vector of p-values. Used for the
adaptive Benjamini-Hochberg method for multiple testing adjustment (chosen
by \code{adjust = "adaptBH"}). Accepts the provided options of the
\code{method} argument of \code{\link[limma]{propTrueNull}}:
\code{"convest"}(default), \code{"lfdr"}, \code{"mean"}, and \code{"hist"}.
Can alternatively be \code{"farco"} for
the "iterative plug-in method" proposed by \cite{Farcomeni (2007)}.}

\item{p_true_null}{proportion of true null hypothesis used for the adaptBH
method. If \code{NULL}, the proportion is computed using the method
defined via \code{true_null_method}.}

\item{verbose}{if \code{TRUE}, progress messages are returned.}
}
\description{
The functions adjusts a vector of p-values for multiple testing
}
